moving some parts from validator to PromptInjectionValidator
